kafka-socket-01/requirements.txt
kafka-python
websocket_client
